users_service/requirements.txt
Flask==2.0.1
PyJWT==2.1.0
Werkzeug==2.0.1
pymongo==3.12.0
Flask-CORS==3.0.10

